split load-data.worker into files
