+++ modules/web/helpers/worker.load-data/lib-dispatch.js
@@ -1,0 +1,5 @@
+// @flow
+
+export default function dispatch(type: string, action: string, ...args: any[]) {
+    self.postMessage([null, 'dispatch', { type, action, args }])
+}
