+++ modules/web/helpers/worker.load-data/lib-log.js
@@ -1,0 +1,4 @@
+// @flow
+
+export default (...args: any[]) =>
+    args.length && console.log('worker:load-data', ...args)
