+++ modules/web/helpers/worker.load-data/types.js
@@ -1,0 +1,3 @@
+// @flow
+
+export type InfoFileTypeEnum = 'courses' | 'areas';
